Changed the contentList sorting in delete_content for listbox
When sorting a copy of a list, the list is sorted to, so I create a new list for deleted content in listbox.
When I delete displayed content, the listbox read the displayed content in reverse alphabetical order, but the deletion loop read them in alphabetical number.
It needed to be sorted so the lbx and the loop read the same value at every iteration